Expand words in line to multiple lines with a function
The usual

:'<,'>s/\%V /\r/

doesn't work to substitute every space within the visual selection
because after changing the first space with a newline the visual
selection marks (`< and `>) no longer indicate the desired region,
since the following text is one line lower than it was at the beginning.
Therefore, we need  a more sophisticated procedure, a function, to
perform the substitution properly.